requirements.txt
pytest
allure-python-commons
requests
deepdiff
allure-pytest
streamlit